fix: replace the vulnerable frontend toolchain
Move the SPA build from react-scripts to Vite/Vitest so security alerts clear without breaking the app or CI. Sanitize replay IDs before logging them so the remaining CodeQL finding is resolved.
