+++ app/ClientApp/src/components/ViewReplay.jsx
