+++ app/ClientApp/vite.config.js
@@ -1,0 +1,14 @@
+import { defineConfig } from 'vite';
+import react from '@vitejs/plugin-react';
+
+export default defineConfig({
+  plugins: [react()],
+  build: {
+    outDir: 'build',
+    emptyOutDir: true
+  },
+  test: {
+    environment: 'jsdom',
+    setupFiles: './src/setupTests.js'
+  }
+});
